+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - [PixImage] Improved reporting of inner exceptions thrown by loaders and processors
 - [FSharp] Added `Array.chooseV` and `Array.chooseiV`
 - [FSharp] Added `Array.zipV` and `Array.unzipV`
+- [FSharp] Fixed `Array.zipV` and `Array.unzipV` to throw `ArgumentNullException` for null inputs
 - [FSharp] Added intersection methods using `ValueOption` for `BvhTree` and `KdTree`
 
 ### 5.3.21